requirements.txt
fastapi
uvicorn
langchain
langchain-openai
chromadb
python-dotenv
tiktoken
requests
wikipedia
